fix(images): make collector metadata portable
